FAB-16544 Fix IT UpdateChannel to match doc
The documentation in UpdateChannel function says that it waits for the
config to commit at all of the peers, but this isn't actually the case.
Instead, it waits for the config to commit at the orderer it's already
committed at.  This change already existed in master, but was apparently
not backported to release-1.4.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>
Change-Id: I83a9f78bba37181b45dd5b3ed4e9af23f3556661